% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors_unordered.R
\name{factors_unordered}
\alias{factors_unordered}
\title{Factor with levels as they appear}
\usage{
factors_unordered(fact)
}
\arguments{
\item{fact}{vector}
}
\value{
factor
}
\description{
Factor with levels as they appear
}
\examples{
factors_unordered(c("a","Hello", "b", "apple"))
}

