% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reordering.R
\name{reordering}
\alias{reordering}
\title{Arrange data frame in descending order based on the factor}
\usage{
reordering(df, a_factor)
}
\arguments{
\item{df}{data frame}

\item{a_factor}{a factor}
}
\value{
data frame
}
\description{
Arrange data frame in descending order based on the factor
}
\examples{
reordering(tbl_df(gapminder), tbl_df(gapminder)$lifeExp)
}

